# メモリを効率良く使う